修正双击退出在ActivityGroup环境下出现KeyEvent事件重发导致双击退出失效的问题

+++ library/src/com/lurencun/android/system/DoubleClickExit.java
@@ -12,26 +12,29 @@ import android.widget.Toast;
  * 双击退出应用
  *
  * @author 桥下一粒砂 (chenyoca@gmail.com)
- * @date   2013-5-8
+ * date   2013-5-8
  */
 public class DoubleClickExit {
 
+    private static final int MIN_CLICK_DELAY = 300;// 300ms
 	private static final String DEFAULT_EXIT_TIP = "再按一次返回键退出应用！";
-	private static final int DEFAULT_EXIT_TIME  = 2000;
+	private static final int DEFAULT_EXIT_TIME  = 2000; // 2s
 
 	private Activity activity;
 	private AtomicBoolean allowToExit = new AtomicBoolean(false);
-	private AtomicBoolean isTaskExcuting = new AtomicBoolean(false);
+	private AtomicBoolean isTaskExecuting = new AtomicBoolean(false);
 
 	private class ResetTask extends TimerTask{
 		@Override
 		public void run() {
 			allowToExit.set(false);
-			isTaskExcuting.set(false);
+			isTaskExecuting.set(false);
 		}
-	};
+	}
 
 	private Timer resetTimer = new Timer("reset-timer", true);
+
+    private long lastClickEventTime = 0;
 
 	/**
 	 * 需要执行退出提示的Activity对象引用 
@@ -49,39 +52,36 @@ public class DoubleClickExit {
 	 * @return false
 	 */
 	public boolean onKeyClick(int keyCode, String tip, int waitTime) {
-		if (keyCode == KeyEvent.KEYCODE_BACK) {
-			if(allowToExit.get()){
-				activity.finish();
-				android.os.Process.killProcess(android.os.Process.myPid());
-			}else{
-				Toast.makeText(activity, tip, Toast.LENGTH_SHORT).show();
-				allowToExit.set(true);
-				// 过一段时间后，恢复重新判定。
-				if( !isTaskExcuting.get() ){
-					isTaskExcuting.set(true);
-					resetTimer.schedule(new ResetTask(), waitTime);
-				}
-			}
+        boolean flag = false;
+        if (keyCode == KeyEvent.KEYCODE_BACK) {
+            long currentClickEventTime = System.currentTimeMillis();
+            final boolean delayFlag = MIN_CLICK_DELAY < currentClickEventTime - lastClickEventTime;
+            if(!delayFlag) return flag;
+            lastClickEventTime = currentClickEventTime;
+            if(allowToExit.get()){
+                activity.finish();
+                android.os.Process.killProcess(android.os.Process.myPid());
+            }else{
+                Toast.makeText(activity, tip, Toast.LENGTH_SHORT).show();
+                allowToExit.set(true);
+                // 过一段时间后，恢复重新判定。
+                if( !isTaskExecuting.get() ){
+                    isTaskExecuting.set(true);
+                    resetTimer.schedule(new ResetTask(), waitTime);
+                }
+                flag = true;
+            }
 
-//			if (!allowToExit.get()) {
-//				allowToExit.set(true);
-//				Toast.makeText(activity, tip, Toast.LENGTH_SHORT).show();
-//				if (!isTaskExcuting.get()) {
-//					
-//				}
-//			} else {
-//				
-//			}
 		}
-		return false;
+		return flag;
 	}
 
 	/**
 	 * 在给定时间内双击某按钮退出程序
-	 * @param keyCode
-	 * @param tip
-	 * @param waitTime
-	 * @return
+	 * @param keyCode 按钮ID
+	 * @param tip 提示文本资源ID
+	 * @param waitTime 双击检测时间
+	 * @return false
 	 */
 	public boolean onKeyClick(int keyCode,int tip,int waitTime){
 		return onKeyClick(keyCode, activity.getResources().getString(tip), waitTime);
